general: add gtk font and cursor + improve query of disk and theme and gpu
It's a little bug and perfomance regression,
where each time disk or theme or gpu are constructed, they query the same thing along all the others again
example:
```
layout = [
        "Theme: $<theme-gtk3.name>",
        "Icons: $<theme-gtk3.icons>"
]
```
It will query the same gtk3 theme name and icons and all others stuff twice.
Thus it will slow down A LOT customfetch

And another problem with them all, is that they have a private member static bool where if true,
it won't query all the infos again. so for example:
```
layout = [
        "Disk (/) free: $<disk(/).free>",
        "Disk (/tmp) free: $<disk(/tmp).total>"
]
```
It will show the query of disk(/) because the bool is true after querying it
and thus it won't query the disk(/tmp)