android app: make the logo type small by default (read below message)
ok listen: the android widget has a major problem, it doesn't wrap lines in any way.
You will have a better experience if you use it on a Tablet, but everyone uses a phone.

I'm currently looking for anyone who would suggest a solution to this, either:
* a way to "disable" wrapping lines like in a terminal OR
* I don't know

Now I made this workaround that basically makes the default ascii-logo-type small, since the android_small.txt logo should fit in every phone.
Please if you know a way to kinda fix this or improve, or else I wasted almost a month of development to the actual program