android app/widget: add default widget configuration value in app
I'm sorry for being away for almost a week.
My original intention was to bring customizable widget update interval in app, togheter with these options,
but the fucking toolbar with `PreferenceFragmentCompat()` doesn't work along as in the normal `Fragment()`,
so now I have wasted 2/3 days just to get that thing work.
shoout to https://stackoverflow.com/questions/64971196/adding-a-toolbar-to-a-preferencescreen-fragment for providing me the solution, tysm :)

Welp, now I have added a new setting page for default values in the configuration of widgets.
Also have added cool animations when clicking the buttons of the app